:art: Refactor ops and fxns to live in models
[Description]
We've figured out the interface for calculating physical phenomina
and performing independent model energy calculations. For this commit
we are putting derivatives and more complex ops in model ops.

We start with a notion of model which can be a basis function used
for the thing we want to predict. Models will live in normal fxn
namespace for lack of a better idea.